Rename namespace to Diagnostics
